Exposing minimal semver API. See #16
