+++ bumpit/core/versions/__init__.py
@@ -1,15 +1,8 @@
-from bumpit.core.versions.semver import SemVer
-from bumpit.core.versions.transformers import IncrementingTransformer, StaticTransformer
+from bumpit.core.versions.semver import next_semantic_version
 
 
 def next_version(strategy, current_version, part, force_value):
     if strategy == "semver":
-        version = SemVer.parse(current_version)
-        if force_value is None:
-            transform = IncrementingTransformer()
-            return str(transform(part, version))
-        else:
-            transform = StaticTransformer()
-            return str(transform(part, version, force_value))
+        return next_semantic_version(current_version, part, force_value)
 
     raise ValueError(f"Unsupported strategy {strategy}")
